Fix .MOD struct layout on LP64, landscape orientation, FORTIFY off
- modfile.h: 'long state' -> int (8 bytes on LP64 broke the 208-byte file layout,
  corrupted triangle reads and crashed level load)
- SDL_HINT_ORIENTATIONS landscape on Android
- disable _FORTIFY_SOURCE for the legacy code paths

Verified on emulator: full flow player creation -> menu -> Castella 1 in-game,
touch input, tutorial, save persistence.

+++ src/modfile.h
@@ -11,8 +11,8 @@ class CD3DEngine;
 
 typedef struct
 {
-	char			bUsed;		// TRUE -> utilisé
-	char			bSelect;	// TRUE -> sélectionné
+	char			bUsed;		// TRUE -> utilisï¿½
+	char			bSelect;	// TRUE -> sï¿½lectionnï¿½
 	D3DVERTEX2		p1;
 	D3DVERTEX2		p2;
 	D3DVERTEX2		p3;
@@ -20,7 +20,7 @@ typedef struct
 	char			texName[20];
 	float			min;
 	float			max;
-	long			state;
+	int				state;		// 'long' is 8 bytes on LP64 (Android) and would break the 208-byte .MOD file layout
 	short			texNum2;
 	short			reserve2;
 	short			reserve3;
